swarms-master-rustified/tests/structs/test_company_rustified.rs
### Compatibility Analysis

The provided Python file appears to be a test suite for a Company class, which manages a collection of Agent instances. The code utilizes pytest for testing and relies on custom classes (`Agent`, `Company`, and `OpenAIChat`) from external modules. The conversion of this code to Rust is viable, but it requires careful consideration of the following aspects:

1.  **External Dependencies**: The Python code uses external libraries like `pytest` and custom modules (`swarm_models`, `swarms.structs.agent`, and `swarms.structs.company`). To maintain compatibility, equivalent Rust libraries and modules must be used or implemented.

2.  **Error Handling**: Rust's error handling mechanism differs from Python's. The Rust code will need to use `Result` and `Option` to handle errors and null values.

3.  **Mocking**: The Python code creates mock objects for testing. In Rust, mocking can be achieved using libraries like `mockall` or by creating custom mock implementations.

4.  **Custom Classes**: The `Agent` and `Company` classes need to be reimplemented in Rust, maintaining their original behavior.

### Rust Conversion

Here is the equivalent Rust code:

```rust
// Conversion Assessment:
// This conversion is viable, as the Python code's logic can be directly translated to Rust.
// However, the code's dependencies on external libraries and custom modules require careful consideration.

// Import Rust libraries
#[cfg(test)]
mod tests {
    use std::collections::HashSet;

    // Define a mock OpenAIChat struct
    pub struct OpenAIChat {
        openai_api_key: String,
        max_tokens: u32,
    }

    impl OpenAIChat {
        pub fn new(openai_api_key: &str, max_tokens: u32) -> Self {
            Self {
                openai_api_key: openai_api_key.to_string(),
                max_tokens,
            }
        }
    }

    // Define an Agent struct
    pub struct Agent {
        llm: OpenAIChat,
        name: String,
    }

    impl Agent {
        pub fn new(llm: OpenAIChat, name: &str) -> Self {
            Self { llm, name: name.to_string() }
        }
    }

    // Define a Company struct
    pub struct Company {
        org_chart: Vec<Vec<Agent>>,
        shared_instructions: String,
        agents: HashSet<Agent>,
    }

    impl Company {
        pub fn new(org_chart: Vec<Vec<Agent>>, shared_instructions: &str) -> Self {
            let mut agents = HashSet::new();
            for row in org_chart {
                for agent in row {
                    agents.insert(agent.clone());
                }
            }
            Self {
                org_chart,
                shared_instructions: shared_instructions.to_string(),
                agents,
            }
        }

        pub fn add(&mut self, agent: Agent) -> Result<(), String> {
            if self.agents.contains(&agent) {
                return Err("Agent already exists".to_string());
            }
            self.agents.insert(agent.clone());
            Ok(())
        }

        pub fn get(&self, name: &str) -> Result<Agent, String> {
            for agent in &self.agents {
                if agent.name == name {
                    return Ok(agent.clone());
                }
            }
            Err("Agent not found".to_string())
        }

        pub fn remove(&mut self, agent: Agent) -> Result<(), String> {
            if !self.agents.contains(&agent) {
                return Err("Agent does not exist".to_string());
            }
            self.agents.remove(&agent);
            Ok(())
        }
    }

    #[test]
    fn test_add_agent() {
        // Create a mock OpenAIChat instance
        let llm = OpenAIChat::new("test_key", 4000);

        // Create mock Agents
        let ceo = Agent::new(llm.clone(), "CEO");
        let dev = Agent::new(llm.clone(), "Developer");
        let va = Agent::new(llm.clone(), "VA");
        let hr = Agent::new(llm.clone(), "HR");
        let shared_instructions = "Listen to your boss";

        // Create a Company instance
        let mut company = Company::new(vec![vec![ceo, dev, va]], shared_instructions);

        // Add the HR agent
        company.add(hr).unwrap();

        // Check if the HR agent exists
        assert!(company.agents.contains(&hr));
    }

    #[test]
    fn test_get_agent() {
        // Create a mock OpenAIChat instance
        let llm = OpenAIChat::new("test_key", 4000);

        // Create mock Agents
        let ceo = Agent::new(llm.clone(), "CEO");
        let dev = Agent::new(llm.clone(), "Developer");
        let va = Agent::new(llm.clone(), "VA");
        let hr = Agent::new(llm.clone(), "HR");
        let shared_instructions = "Listen to your boss";

        // Create a Company instance
        let mut company = Company::new(vec![vec![ceo, dev, va]], shared_instructions);

        // Add the HR agent
        company.add(hr).unwrap();

        // Get the HR agent
        let retrieved_agent = company.get("HR").unwrap();

        // Check if the retrieved agent matches the expected agent
        assert_eq!(retrieved_agent.name, hr.name);
    }

    #[test]
    fn test_remove_agent() {
        // Create a mock OpenAIChat instance
        let llm = OpenAIChat::new("test_key", 4000);

        // Create mock Agents
        let ceo = Agent::new(llm.clone(), "CEO");
        let dev = Agent::new(llm.clone(), "Developer");
        let va = Agent::new(llm.clone(), "VA");
        let hr = Agent::new(llm.clone(), "HR");
        let shared_instructions = "Listen to your boss";

        // Create a Company instance
        let mut company = Company::new(vec![vec![ceo, dev, va]], shared_instructions);

        // Add the HR agent
        company.add(hr).unwrap();

        // Remove the HR agent
        company.remove(hr).unwrap();

        // Check if the HR agent no longer exists
        assert!(!company.agents.contains(&hr));
    }

    #[test]
    fn test_add_existing_agent() {
        // Create a mock OpenAIChat instance
        let llm = OpenAIChat::new("test_key", 4000);

        // Create mock Agents
        let ceo = Agent::new(llm.clone(), "CEO");
        let dev = Agent::new(llm.clone(), "Developer");
        let va = Agent::new(llm.clone(), "VA");
        let hr = Agent::new(llm.clone(), "HR");
        let shared_instructions = "Listen to your boss";

        // Create a Company instance
        let mut company = Company::new(vec![vec![ceo, dev, va]], shared_instructions);

        // Add the HR agent
        company.add(hr).unwrap();

        // Attempt to add the HR agent again
        let result = company.add(hr);
        assert!(result.is_err());
    }

    #[test]
    fn test_get_nonexistent_agent() {
        // Create a mock OpenAIChat instance
        let llm = OpenAIChat::new("test_key", 4000);

        // Create mock Agents
        let ceo = Agent::new(llm.clone(), "CEO");
        let dev = Agent::new(llm.clone(), "Developer");
        let va = Agent::new(llm.clone(), "VA");
        let shared_instructions = "Listen to your boss";

        // Create a Company instance
        let company = Company::new(vec![vec![ceo, dev, va]], shared_instructions);

        // Attempt to get a nonexistent agent
        let result = company.get("Nonexistent");
        assert!(result.is_err());
    }

    #[test]
    fn test_remove_nonexistent_agent() {
        // Create a mock OpenAIChat instance
        let llm = OpenAIChat::new("test_key", 4000);

        // Create mock Agents
        let ceo = Agent::new(llm.clone(), "CEO");
        let dev = Agent::new(llm.clone(), "Developer");
        let va = Agent::new(llm.clone(), "VA");
        let hr = Agent::new(llm.clone(), "HR");
        let shared_instructions = "Listen to your boss";

        // Create a Company instance
        let mut company = Company::new(vec![vec![ceo, dev, va]], shared_instructions);

        // Attempt to remove a nonexistent agent
        let result = company.remove(hr);
        assert!(result.is_err());
    }
}
```

In the Rust code:

*   **Mocking**: We've implemented mock structs for `OpenAIChat` and `Agent` to replace the Python mocking.
*   **Error Handling**: Rust's `Result` type is used to handle errors in the `add`, `get`, and `remove` methods.
*   **Custom Classes**: Equivalent Rust implementations of the `Agent` and `Company` classes are provided, maintaining their original behavior.
*   **Testing**: We've used the `#[test]` attribute to define test functions for the Rust code, replacing the Python `pytest` library.

The conversion process has allowed us to maintain the original functionality of the Python code while taking advantage of Rust's safety features and performance. However, differences in error handling, mocking, and custom classes required careful consideration during the conversion process.